move tabs to navigation
