+++ src/lib/components/Navigation/Tabs/index.ts
