+++ src/lib/components/Navigation/index.ts
@@ -1,3 +1,4 @@
 import Navbar from './Navbar';
+import Tabs from './Tabs';
 
-export { Navbar };
+export { Navbar, Tabs };
